separate class-map-list
