+++ catalystwan/endpoints/configuration/policy/list/class_map.py
@@ -5,17 +5,9 @@ from uuid import UUID
 
 from catalystwan.endpoints import APIEndpoints, delete, get, post, put
 from catalystwan.endpoints.configuration.policy.abstractions import PolicyListEndpoints
-from catalystwan.models.policy.lists import ClassMapList
-from catalystwan.models.policy.policy_list import InfoTag, PolicyListId, PolicyListInfo, PolicyListPreview
+from catalystwan.models.policy.list.class_map import ClassMapList, ClassMapListEditPayload, ClassMapListInfo
+from catalystwan.models.policy.policy_list import InfoTag, PolicyListId, PolicyListPreview
 from catalystwan.typed_list import DataSequence
-
-
-class ClassMapListEditPayload(ClassMapList, PolicyListId):
-    pass
-
-
-class ClassMapListInfo(ClassMapList, PolicyListInfo):
-    pass
 
 
 class ConfigurationPolicyForwardingClassList(APIEndpoints, PolicyListEndpoints):
